Switch from `uint` to `long` just in case

+++ src/Synology.Api.Client/Apis/FileStation/List/Models/FileStationListShareOwner.cs
@@ -4,11 +4,11 @@ namespace Synology.Api.Client.Apis.FileStation.List.Models
 {
     public class FileStationListShareOwner
     {
-        public uint Gid { get; set; }
+        public long Gid { get; set; }
 
         public string Group { get; set; }
 
-        public uint Uid { get; set; }
+        public long Uid { get; set; }
 
         public string User { get; set; }
     }
